requirements_base.txt
Pillow>=3
CairoSVG>=1.0
django-allauth>=0.36
django-bootstrap-form>=3.2
django-compressor>=1.5
django-ipware>=1.1
django-leaflet
django-modeltranslation>=0.13-beta1
django-recaptcha>=1.2.1
djangorestframework>=3.8
djangorestframework-gis
django-cors-headers>=2.0.2
markdown2>=2.3
psycopg2>=2.6
requests>=2.11
sorl-thumbnail>=12.3
python3-memcached>=1.51
psutil>=3.4
